third_party/blink/web_tests/fast/html/imports/import-normalize-resolve-4-expected.txt

CONSOLE WARNING: line 11: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE MESSAGE: line 10: e01
CONSOLE MESSAGE: line 10: e02
CONSOLE MESSAGE: line 10: e03
CONSOLE MESSAGE: line 10: e04
CONSOLE MESSAGE: line 10: e05
CONSOLE MESSAGE: line 10: e06
CONSOLE MESSAGE: line 10: e07
CONSOLE MESSAGE: line 10: e08
CONSOLE MESSAGE: line 10: e09
CONSOLE MESSAGE: line 10: e10
CONSOLE MESSAGE: line 10: e11
CONSOLE MESSAGE: line 10: e12
CONSOLE MESSAGE: line 10: e13
CONSOLE MESSAGE: line 10: e14
CONSOLE MESSAGE: line 10: e15
CONSOLE MESSAGE: line 10: e16
CONSOLE MESSAGE: line 10: e17
CONSOLE MESSAGE: line 10: e18
CONSOLE MESSAGE: line 10: e19
CONSOLE MESSAGE: line 10: e20
CONSOLE MESSAGE: line 10: e21
